requirements.txt
selenium==4.15.0
pytest==7.4.3
pytest-html==4.1.1
allure-pytest==2.13.2
webdriver-manager==4.0.1
pytest-xdist==3.5.0
pytest-ordering==0.6
python-dotenv==1.0.0